refactor: engine handlers
