+++ pkg/engine/generation.go
@@ -2,10 +2,8 @@ package engine
 
 import (
 	"context"
-	"time"
 
 	"github.com/go-logr/logr"
-	kyvernov1beta1 "github.com/kyverno/kyverno/api/kyverno/v1beta1"
 	"github.com/kyverno/kyverno/pkg/autogen"
 	engineapi "github.com/kyverno/kyverno/pkg/engine/api"
 	"github.com/kyverno/kyverno/pkg/engine/internal"
@@ -16,16 +14,6 @@ func (e *engine) generateResponse(
 	ctx context.Context,
 	logger logr.Logger,
 	policyContext engineapi.PolicyContext,
-	gr kyvernov1beta1.UpdateRequest,
-) engineapi.PolicyResponse {
-	return e.filterGenerateRules(policyContext, logger, gr.Spec.Policy, time.Now())
-}
-
-func (e *engine) filterGenerateRules(
-	policyContext engineapi.PolicyContext,
-	logger logr.Logger,
-	policyNameKey string,
-	startTime time.Time,
 ) engineapi.PolicyResponse {
 	resp := engineapi.NewPolicyResponse()
 	for _, rule := range autogen.ComputeRules(policyContext.Policy()) {
