+++ pkg/engine/image_verify.go
@@ -9,20 +9,20 @@ import (
 	"github.com/kyverno/kyverno/pkg/autogen"
 	engineapi "github.com/kyverno/kyverno/pkg/engine/api"
 	"github.com/kyverno/kyverno/pkg/engine/handlers"
+	"github.com/kyverno/kyverno/pkg/engine/handlers/mutation"
 	"github.com/kyverno/kyverno/pkg/engine/internal"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
-// mutate performs mutation. Overlay first and then mutation patches
-func (e *engine) mutate(
+func (e *engine) verifyAndPatchImages(
 	ctx context.Context,
 	logger logr.Logger,
 	policyContext engineapi.PolicyContext,
-) (engineapi.PolicyResponse, unstructured.Unstructured) {
+) (engineapi.PolicyResponse, engineapi.ImageVerificationMetadata) {
 	resp := engineapi.NewPolicyResponse()
 	policy := policyContext.Policy()
 	matchedResource := policyContext.NewResource()
 	applyRules := policy.GetSpec().GetApplyRules()
+	ivm := engineapi.ImageVerificationMetadata{}
 
 	policyContext.JSONContext().Checkpoint()
 	defer policyContext.JSONContext().Restore()
@@ -31,13 +31,17 @@ func (e *engine) mutate(
 		startTime := time.Now()
 		logger := internal.LoggerWithRule(logger, rule)
 		handlerFactory := func() (handlers.Handler, error) {
-			if !rule.HasMutate() {
+			if !rule.HasVerifyImages() {
 				return nil, nil
 			}
-			if !policyContext.AdmissionOperation() && rule.IsMutateExisting() {
-				return e.mutateExistingHandler, nil
-			}
-			return e.mutateResourceHandler, nil
+			return mutation.NewMutateImageHandler(
+				policyContext,
+				matchedResource,
+				rule,
+				e.configuration,
+				e.rclient,
+				&ivm,
+			)
 		}
 		resource, ruleResp := e.invokeRuleHandler(
 			ctx,
@@ -46,7 +50,7 @@ func (e *engine) mutate(
 			policyContext,
 			matchedResource,
 			rule,
-			engineapi.Mutation,
+			engineapi.ImageVerify,
 		)
 		matchedResource = resource
 		for _, ruleResp := range ruleResp {
@@ -58,5 +62,6 @@ func (e *engine) mutate(
 			break
 		}
 	}
-	return resp, matchedResource
+	// TODO: i doesn't make sense to not return the patched resource here
+	return resp, ivm
 }
